chore(test): move dev related options to rolldown_dev_common (#6667)
I added `rolldown_dev_common`​ because types in `rolldown_dev_common`​ only relies on `rolldown_common`​. To emphasize the relationship, I used a new crate. Also splitting to more crates to friendly to compiling in paralell.
