+++ crates/rolldown/src/dev/mod.rs
@@ -4,15 +4,15 @@ pub mod build_state_machine;
 pub mod building_task;
 pub mod dev_context;
 pub mod dev_engine;
-pub mod dev_options;
-pub mod rebuild_strategy;
 pub mod types;
 pub mod watcher_event_handler;
 
 use std::sync::Arc;
 
-pub use dev_options::{DevOptions, NormalizedDevOptions, OnHmrUpdatesCallback};
-pub use rebuild_strategy::RebuildStrategy;
+pub use rolldown_dev_common::types::{
+  BundleOutput, DevOptions, DevWatchOptions, NormalizedDevOptions, OnHmrUpdatesCallback,
+  OnOutputCallback, RebuildStrategy, SharedNormalizedDevOptions, normalize_dev_options,
+};
 use rolldown_utils::dashmap::FxDashMap;
 
 use crate::dev::types::client_session::ClientSession;
