+++ crates/rolldown/src/types/bundle_output.rs
@@ -1,8 +1,1 @@
-use rolldown_common::Output;
-use rolldown_error::BuildDiagnostic;
-
-#[derive(Default)]
-pub struct BundleOutput {
-  pub warnings: Vec<BuildDiagnostic>,
-  pub assets: Vec<Output>,
-}
+pub use rolldown_dev_common::types::BundleOutput;
