+++ crates/rolldown_binding/src/binding_dev_engine.rs
@@ -1,6 +1,6 @@
 use napi_derive::napi;
 use rolldown::dev::dev_context::BuildProcessFuture;
-use rolldown::dev::{OnHmrUpdatesCallback, dev_options::OnOutputCallback};
+use rolldown::dev::{OnHmrUpdatesCallback, OnOutputCallback};
 use std::path::PathBuf;
 use std::sync::Arc;
 
@@ -106,7 +106,7 @@ impl BindingDevEngine {
       || compare_contents_for_polling.is_some()
       || debounce_tick_rate.is_some()
     {
-      Some(rolldown::dev::dev_options::DevWatchOptions {
+      Some(rolldown::dev::DevWatchOptions {
         disable_watcher: None,
         skip_write,
         use_polling,
@@ -120,7 +120,7 @@ impl BindingDevEngine {
       None
     };
 
-    let rolldown_dev_options = rolldown::dev::dev_options::DevOptions {
+    let rolldown_dev_options = rolldown::dev::DevOptions {
       on_hmr_updates,
       on_output,
       rebuild_strategy,
