+++ crates/rolldown_dev_common/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "rolldown_dev_common"
+version = "0.1.0"
+publish = true
+description = "Common types for rolldown dev mode"
+
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+repository.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+doctest = false
+
+[lints]
+workspace = true
+
+[dependencies]
+derive_more = { workspace = true }
+rolldown_common = { workspace = true }
+rolldown_error = { workspace = true }
