+++ crates/rolldown_dev_common/src/types/bundle_output.rs
@@ -1,0 +1,8 @@
+use rolldown_common::Output;
+use rolldown_error::BuildDiagnostic;
+
+#[derive(Default)]
+pub struct BundleOutput {
+  pub warnings: Vec<BuildDiagnostic>,
+  pub assets: Vec<Output>,
+}
