+++ crates/rolldown_testing/src/integration_test.rs
@@ -8,8 +8,7 @@ use std::{
 };
 
 use anyhow::Context;
-use rolldown::dev::DevOptions;
-use rolldown::dev::dev_options::DevWatchOptions;
+use rolldown::dev::{DevOptions, DevWatchOptions};
 use rolldown::{
   BundleOutput, Bundler, BundlerOptions, IsExternal, OutputFormat, Platform, SourceMapType,
   plugin::__inner::SharedPluginable,
